Perubahan nama file migrasi
Menyelesaikan isu menjalankan migrasi menggunakan artisan (pendefinisian
foreign key pada table post_jobs dilakukan setelah migrasi table companies)

+++ database/migrations/2023_01_18_183449_create_post_jobs_table.php
@@ -15,7 +15,7 @@ class CreatePostJobsTable extends Migration
     {
         Schema::create('post_jobs', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('company_id')->nullable();
+            $table->unsignedBigInteger('company_id')->nullable();
             $table->string('title', 128)->index();
             $table->string('desc', 4096);
             $table->string('benefit', 512);
